implement some SIM IO (phone # and ICCID)
Some of it is not completely correct, but it worked in my tests.

It could be nicer to just patch into the telephony framework with
a SimFileHandler that drectly uses oFono data rather than fake SIM
I/O, but for now I think it would be good to be a replacement
for RIL.java only.