fix "random" dbus disconnect
Android's LocalSocket class doesn't handle timeouts well, so when they
happen you just get a IOException with the message "Try again"
(errno EAGAIN). This was blowing up MessageReader, which would cascade
down to AbstractConnection, which would disconnect.

The somewhat hacky workaround is to make a wrapper InputStream class
that looks for "Try again" and reports it as a SocketTimeout rather
than IOException.