refactor: wrapper class that handles the async parts
Should remove repetitive code, and chances for errors like doing stuff on the
wrong thread, or an exception that crashes the phone process or never responds
to the caller after the error occurrred.

One downside is that now *everything* is shuttled to another thread, even the
trivial ones that just return some props. This can be addressed later by something
like a @SafeForMainThread annotation.