improve datacall implementation
Bugfixes, work with oFono driver other than rilmodem.

* Notify VoiceNetworkStateRegistrants when we attach to the data
network -- seems to be the way to get the framework to re-poll data
registration state and start any datacalls it wants to. This should
fix the problems where the framework wouldn't even try to start a
datacall.

* Fall back on voice radio technology value if oFono doesn't provide
data technology. Do it by dependency injection of a limited interface
to minimize coupling.

* Return even inactive datacalls (not sure that it matters much).

* Return the netmask, in e.g. "/30" notation as expected by framework
code.

* Setup/teardown the network interface, using
INetworkManagementService.  I guess the oFono rilmodem plugin was
doing this for us as a side-effect. I found the network management
service by searching for Java code that could do that stuff, but we
could switch to something else if needed.

Currently relying on oFono to give us the interface name (patched my
gobidev plugin to specify rmnet0), but I wonder if interface assignment
is something better done outside oFono. Does the modem care which
interface is used for which context? It seems weird, in this usecase
at least, for oFono to give us an interface, if it has no relation to
the modem. Could investigate rild's message "Interface rmnet1 will be
used for CID(2)" to see how it decides.

Some errors I was seeing early on may have come from oFono not having
an interface assigned, so if we do leave that unset on their end
maybe there's something to work around.