(un)load modules in response to oFono's interfaces
* restructure inter-module dependencies to support this

* dispatch all signals through RilOfono class so we don't have to deal
with registering/unregistering signals as modules come and go

* store module references in RilOfono (rather than RilWrapper) with
internal interface types, so we can dispatch signals

* remove _unknown registration state (not a real oFono string),
use "none" instead

* remove VoiceRadioTechnologyGetter, because the code using it ended
up needing more netreg interaction, to the point we may as well just
use the oFono stuff directly

* allow suppressing propchange logs, because two modules are now
tracking netreg props

* returning RADIO_NOT_AVAILABLE in the RilWrapper, when modules are
null, seemed to prevent the framework from coming online, so just use
GENERIC_FAILURE with a modified log message

* work around issue where ConnManager never reports Attached. It seems
to be a new issue, I guess due to timing differences with this new
implementation. The workaround is to watch NetworkRegistration for
Status changes, and reload ConnectionManager's props in response.